feat: implement skeletons for RabbitMQ

+++ board-service/internal/models/board_model.go
@@ -3,10 +3,10 @@ package model
 import models "github.com/sm888sm/halten-backend/models"
 
 type Pagination struct {
-	CurrentPage  int
-	TotalPages   int
-	ItemsPerPage int
-	TotalItems   int
+	CurrentPage  uint64
+	TotalPages   uint64
+	ItemsPerPage uint64
+	TotalItems   uint64
 	HasMore      bool
 }
